--- autolift-cats/src/main/scala/autolift/cats/Implicits.scala
@@ -1,6 +1,0 @@
-package autolift.cats
-
-trait Implicits extends LiftedMapImplicits
-	with LiftedApImplicits
-	with LiftedFlatMapImplicits
-	with LiftedMergeWithImplicits
