+++ autolift-cats/src/main/scala/autolift/cats/LiftA*.scala
@@ -1,0 +1,44 @@
+package autolift.cats
+
+import autolift._
+
+trait `LiftA*Package` extends LiftA2Context
+  with LiftA3Context
+  with LiftA4Context
+  with LiftA5Context
+  with LiftA6Context
+  with LiftA7Context
+  with LiftA8Context
+  with LiftA9Context
+  with LiftA10Context
+  with LiftA11Context
+  with LiftA12Context
+  with LiftA13Context
+  with LiftA14Context
+  with LiftA15Context
+  with LiftA16Context
+  with LiftA17Context
+  with LiftA18Context
+  with LiftA19Context
+  with LiftA20Context
+  with LiftA21Context
+  with LiftA2Reexport
+  with LiftA3Reexport
+  with LiftA4Reexport
+  with LiftA5Reexport
+  with LiftA6Reexport
+  with LiftA7Reexport
+  with LiftA8Reexport
+  with LiftA9Reexport
+  with LiftA10Reexport
+  with LiftA11Reexport
+  with LiftA12Reexport
+  with LiftA13Reexport
+  with LiftA14Reexport
+  with LiftA15Reexport
+  with LiftA16Reexport
+  with LiftA17Reexport
+  with LiftA18Reexport
+  with LiftA19Reexport
+  with LiftA20Reexport
+  with LiftA21Reexport
