+++ autolift-cats/src/main/scala/autolift/cats/package.scala
@@ -5,10 +5,12 @@ object functor extends LiftMapPackage
 object applicative extends LiftApPackage
   with LiftMergePackage
   with LiftMergeWithPackage
+  with `LiftA*Package`
 
 object monad extends LiftFlatMapPackage
   with LiftFlattenPackage
   with LiftFilterPackage
+  with `LiftM*Package`
 
 object fold extends LiftFoldPackage
   with LiftFoldLeftPackage
@@ -17,5 +19,5 @@ object fold extends LiftFoldPackage
   with LiftExistsPackage
   with LiftForAllPackage
 
-//object traverse extends LiftSequencePackage
-//  with LiftTraversePackage
+object traverse extends LiftSequencePackage
+  with LiftTraversePackage
